+++ src/app/questions/new/page.tsx
@@ -1,0 +1,9 @@
+import { CreateQuestionCard } from '~/components/domains/Question/CreateQuestionCard';
+
+export default function Page() {
+  return (
+    <div className="flex justify-center pt-[24px]">
+      <CreateQuestionCard />
+    </div>
+  );
+}
