+++ src/components/domains/Question/CreateQuestionCard/index.ts
@@ -1,0 +1,1 @@
+export { CreateQuestionCard } from './CreateQuestionCard';
