--- src/components/domains/Question/CreateQuestionPaper/index.ts
@@ -1,1 +1,0 @@
-export { CreateQuestionPaper } from './CreateQuestionPaper';
